fix to my merging operation failed

+++ megaavr/cores/dxcore/api/PluggableUSB.cpp
@@ -82,7 +82,7 @@ bool PluggableUSB_::plug(PluggableUSBModule *node) {
   node->pluggedEndpoint = lastEp;
   lastIf += node->numInterfaces;
   for (uint8_t i = 0; i < node->numEndpoints; i++) {
-    *(unsigned int *)(epBuffer(lastEp)) = node->endpointType[i];
+    *(uint8_t *)(epBuffer(lastEp)) = node->endpointType[i];
     lastEp++;
   }
   return true;
